陈港5月5日上传的代码

+++ src/main/java/ltd/newbee/mall/dao/NewBeeMallGoodsMapper.java
@@ -75,7 +75,7 @@ public interface NewBeeMallGoodsMapper {
 //参考人数
     boolean insertHelpNum(HelpNum goodsReviewaHelpNum);
     boolean updateReviewNum(HelpNum goodsReviewaHelpNum);
-    long getGoodsReviewHelpNum(long reviewId);
+    long getGoodsReviewHelpNum(int reviewId);
 
 }
